Fix IAR/Keil build warnings and errors

+++ Libraries/PeriphDrivers/Source/TMR/tmr_ai85.c
@@ -27,7 +27,7 @@
 int MXC_TMR_Init(mxc_tmr_regs_t *tmr, mxc_tmr_cfg_t *cfg, bool init_pins)
 {
     int tmr_id = MXC_TMR_GET_IDX(tmr);
-    uint8_t clockSource = MXC_TMR_CLK0;
+    int clockSource = MXC_TMR_CLK0;
 
     if (cfg == NULL) {
         return E_NULL_PTR;
