+++ Libraries/PeriphDrivers/Source/TMR/tmr_me12.c
@@ -26,7 +26,7 @@
 int MXC_TMR_Init(mxc_tmr_regs_t *tmr, mxc_tmr_cfg_t *cfg, bool init_pins, sys_map_t pin_sel)
 {
     uint8_t tmr_id;
-    uint8_t clockSource = MXC_TMR_CLK0;
+    int clockSource = MXC_TMR_CLK0;
 
     if (cfg == NULL) {
         return E_NULL_PTR;
